cleaning + scanner parameter tweak
- move scanrange to parameter
- now mainFrame (paramter+scan) accept drag drop instead of just parameter (uglier in my opinion but more effective)
- add comments for possible feature to implement in treewidget
- remove text added by accident in about.rst
- rename/move functions for lisibility
