+++ docs/about.rst
@@ -15,7 +15,7 @@ Thanks to Maxime, Giuseppe and Guilhem for their contributions.
 
 **You find this package useful?** We would be really grateful if you could help us to improve its visibility ! You can:
 
-* Add a star on the `GitHub page of this project <https://github.com/autolab-project/autolab>`_ of the C2N-CNRS laboratory
+* Add a star on the `GitHub page of this project <https://github.com/autolab-project/autolab>`_
 * Spread the word around you
 * Mention this package in your research publications
